switch rqlite to using database/sql under the hood
Really this is only a partial switch; it is monumentally easier to use
database/sql/driver to export the data, but in order to support multiple
transactions in the next commit, we want to take advantage of the standard
library's connection pooling, etc. Thus, we make this switch.

Additionally, we _have_ to have something like rawConn() (although its
semantics are confusing, see the next patch r.e. in-memory sqlite3 dbs) in
order to make this work, since sqlite3's backup functionality is only
exposed at that level, which we use for snapshotting.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>