test: really test the view insert
Here, the test was slightly wrong, and the way the transaction supporting
code changes behavior slightly exposed this: the last insert id was 3,
which was the rowid of the last actual insert (views don't have insert
ids). Instead, let's do an insert and then a select to check this, and
actually check the right id.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>