store test: clean up after ourselves after each test
This is sort of painful, but it's because we had to switch to using the
same in-memory database due to the way golang's sql driver works. I think
the Right Fix here is really to just stop using in-memory databases
alltogether, and we can get rid of this and Destroy() entirely.

Signed-off-by: Tycho Andersen <tycho.andersen@canonical.com>