Add statement timings support to CLI (#414)
Currently the CLI doesn’t display the statement timings after executing an SQL query. This change adds a new `.timer on|off` command to enable/disable this feature. When enabled, the CLI will append the `timings` URL query parameter to all query/execute requests sent to the API. When a response is received, the execution time is shown below the result.

Resolves: #317