Added go-hclog support and formatted log messages to conform.
hclog uses a weird format for their log messages, rather than doing a
string format on the back end it takes the args for the log message
and prints them as alternating key/values. This replaces the current
logging implementation with hclog (so that the logging interface can be
passed directly down to the underlying raft implementation) and does its
best to make log messages conform to hclogs format.