Intepret expected results as dict and comp
This was done to migrate from Python2 to Python3, which changed the
order of keys in the dict. This was always fragile, and now dictionaries
are compared instead.

Most of this migration was done using sed, which is a great tool. The
command issued was:

sed 's/self\.assertEqual(str(j), "\(.\+\)")/self\.assertEqual(j, d_("\1"))/'