Remove unneeded code now
Still not sure if I should use "raft" as a protocol or not. It's
clearer, but makes notify logic more complicated perhaps. The issue is
that DNS will not have the protocol set. In that case you want to at
least try the Raft protocol. Therefore you have a case where even if
protocol is not set, Raft will be tried. If that's the case, why not
just write the code so that if proto is not set, then Raft should be
tried at some point?