Hook in node client to respond to raft:// joins
Test_SingleJoinOKviaRaft is failing, seems to be hanging. It's not clear
if the Cluster Service object responds properly to a HTTP request, it
should error out and return an error, but that doesn't seem to be
happening. However, even when I comment that code out, the test
node-join is still hanging. Not clear why.