Don't insert records if already in database
When running on Windows this test often encounters ErrLeadershipLost, at
least in the AppVeyor CI environment. This means that INSERT may, or may
not, be actually inserted into the SQLite database. As a result change
this test to use explicit primary keys, so that if the records were
actually inserted, the duplicate records will not be inserted (by also
using the SQLite keyword IGNORE).

See https://github.com/rqlite/rqlite/issues/1123 for more information.