Reduce Raft threshold to 2048
This can now go lower, since Raft snapshotting is more efficient. The
upside is that few logs will need be replayed when a node restarts,
which will result in shorter startup times. We may be able to go even
lower -- Consul defaults to 1024.

The downside is more Snapshotting, which pauses the system during that
time. But the pause times should be very short.