If a load is in progress, block Snapshotting
With chunked loads, a snapshot will delete "LOAD_CHUNK" entries from the
Raft log. However those entries may have not been committed to the
database yet, it won't happen until the last chunk is received. This
would be OK, unless the system restarted before the last chunk and then
the those entries would be lost, but the database not updated -- or
worse, updated with an incomplete load operation (which would result in
a invalid SQLite file).