No checkpoint on close (#1877)
As part of the 9.0 design, rqlite must not checkpoint any database when it's closed, as that would put the primary SQLite database in an invalid state -- WAL files would be checkpointed even though a Snapshot had not occurred. This gets this change into master, reducing the changes that will arrive due to other 9.0 changes.