Remove persisted HWM
There is no need to do this. The state is effectively stored by looking
at the state of the FIFO queue. The persisted highwater mark is the
first key present in the FIFO when the system wakes up, or if it's zero
that means the system is starting, or the queue is empty after being
pruned. And if that's the case there no events will be retransmitted,
which was the point of the persisting the HWM.